new variable "YearsSinceTurnover"
Carla and I were chatting about DV and what we thought that would be, lol.   In that conversation we determined that we really needed another variable that was the time since turnover.  Started created in R, but got confused, so went back to old trusty Altreyx and created a variable called "YearsSinceTurnover" which is today minus last sale date, formatted as a decimal year